chore: ajustado namespaces
