+++ app/Models/Training.php
@@ -2,8 +2,8 @@
 
 namespace App\Models;
 
-use App\Notifications\TrainingNotification;
-use App\Notifications\NotificationConfirmationTrainingNotification;
+use App\Notifications\Training\TrainingNotification;
+use App\Notifications\Training\NotificationConfirmationTrainingNotification;
 use App\Rules\RelationshipSpecificFundamental;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
